refactor: separate store into individual files based on type
